ref: reorganize libs
