+++ libs/ci-config/package.json
@@ -1,4 +1,4 @@
 {
-  "name": "@cypress-dx/ci-config",
+  "name": "@cypress/ci-config",
   "version": "0.0.1"
 }
