+++ libs/ci-config/src/lib/config.ts
@@ -1,3 +1,5 @@
+import { generateCircleCI, generateGitHubActions } from "./providers";
+
 export type CIProvider = 'CircleCI' | 'GitHub Actions';
 export type PackageManager = 'npm' | 'yarn';
 export type CypressTestTypes = 'e2e' | 'component' | 'both';
@@ -12,4 +14,17 @@ export interface GenerateConfig {
   triggerType: CIOnType;
   triggerBranches: string[];
   recordKey: boolean;
+}
+
+export function generate(config: GenerateConfig) {
+  switch (config.provider) {
+    case 'CircleCI':
+      return generateCircleCI(config);
+
+    case 'GitHub Actions':
+      return generateGitHubActions(config);
+
+    default:
+      break;
+  }
 }
