+++ libs/ci-config/src/lib/providers/index.ts
@@ -1,0 +1,2 @@
+export { generateGitHubActions } from './github-actions'
+export { generateCircleCI } from './circle-ci'
