--- libs/ci-generator/README.md
@@ -1,7 +1,0 @@
-# ci-generator
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test ci-generator` to execute the unit tests via [Jest](https://jestjs.io).
