--- libs/ci-generator/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@cypress-dx/ci-generator",
-  "version": "0.0.1"
-}
