--- libs/ci-generator/src/index.ts
@@ -1,1 +1,0 @@
-export { ciGenerate } from './lib/ci-generate';
