+++ libs/cy-scaffold/tsconfig.lib.json
@@ -6,5 +6,5 @@
     "types": ["node"]
   },
   "exclude": ["**/*.spec.ts", "**/*.test.ts"],
-  "include": ["**/*.ts", "src/lib/cli.js"]
+  "include": ["**/*.ts", "src/lib/cli.js"],
 }
