+++ libs/cy-scaffold/tsconfig.spec.json
@@ -3,18 +3,12 @@
   "compilerOptions": {
     "outDir": "../../dist/out-tsc",
     "module": "commonjs",
-    "types": ["jest", "node"]
+    "types": ["vitest/globals", "node"]
   },
   "include": [
-    "jest.config.ts",
-    "**/*.test.ts",
+    "vitest.config.ts",
     "**/*.spec.ts",
-    "**/*.test.tsx",
-    "**/*.spec.tsx",
-    "**/*.test.js",
     "**/*.spec.js",
-    "**/*.test.jsx",
-    "**/*.spec.jsx",
     "**/*.d.ts"
   ]
 }
