--- libs/providers/circle-ci/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@cypress-dx/providers-circle-ci",
-  "version": "0.0.1"
-}
