--- libs/providers/github-actions/jest.config.js
@@ -1,14 +1,0 @@
-module.exports = {
-  displayName: 'providers-github-actions',
-  preset: '../../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  transform: {
-    '^.+\\.[tj]sx?$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../../coverage/libs/providers/github-actions',
-};
