--- libs/providers/github-actions/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "@cypress-dx/providers-github-actions",
-  "version": "0.0.1"
-}
