--- libs/providers/github-actions/src/index.ts
@@ -1,1 +1,0 @@
-export { generateGitHubActions } from './lib/ci-github-actions';
