+++ nx.json
@@ -1,5 +1,5 @@
 {
-  "npmScope": "cypress-dx",
+  "npmScope": "@cypress",
   "affected": {
     "defaultBase": "master"
   },
