+++ tsconfig.base.json
@@ -15,15 +15,8 @@
     "skipDefaultLibCheck": true,
     "baseUrl": ".",
     "paths": {
-      "@cypress-dx/ci-config": ["libs/ci-config/src/index.ts"],
-      "@cypress-dx/ci-generator": ["libs/ci-generator/src/index.ts"],
-      "@cypress-dx/cy-scaffold": ["libs/cy-scaffold/src/index.ts"],
-      "@cypress-dx/providers-circle-ci": [
-        "libs/providers/circle-ci/src/index.ts"
-      ],
-      "@cypress-dx/providers-github-actions": [
-        "libs/providers/github-actions/src/index.ts"
-      ]
+      "@cypress/ci-config": ["libs/ci-config/src/index.ts"],
+      "@cypress/cy-scaffold": ["libs/cy-scaffold/src/index.ts"]
     }
   },
   "exclude": ["node_modules", "tmp"]
